Adds ConsoleLog and tests in a support matrix
This adds `wapc.ConsoleLog` and stubs in the first end-to-end test for
guest integration using wazero.

This adds CI, reverse engineering a support matrix by the min and max
supported Go versions based on latest TinyGo.

Note: This is wider than Go's policy as Go no longer supports 1.16 which
is the version required by TinyGo 0.17.0.

Signed-off-by: Adrian Cole <adrian@tetrate.io>